20210826座標型別變更

+++ Models/EmployeeTrip2Record.cs
@@ -11,8 +11,8 @@ namespace People_errand_api.Models
         public int GroupId { get; set; }
         public string HashAccount { get; set; }
         public int Trip2TypeId { get; set; }
-        public double CoordinateX { get; set; }
-        public double CoordinateY { get; set; }
+        public decimal CoordinateX { get; set; }
+        public decimal CoordinateY { get; set; }
         public DateTime CreatedTime { get; set; }
 
         public virtual Employee HashAccountNavigation { get; set; }
